backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1

# Google OAuth
google-auth==2.37.0
google-auth-oauthlib==1.2.1

# Gemini AI
google-genai==1.0.0

# WebSocket
websockets==14.1

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
aiosqlite==0.20.0

# Dev
ruff==0.8.6
